requirements.txt
torch
torchvision
torchaudio

transformers
sentencepiece

pillow
pandas
numpy
scikit-learn

tqdm

ollama
matplotlib
json